Add a hello world module.
This is our first module that we can build using the containerized
kernel builder image. This commit updates the kbuild.sh script used by the
kernel builder image to prompt the user as to whether they want to build
out-of-tree modules or not. If the answer is yes, the builder image
will build the modules under embedded_linux_drivers/modules leaving the
*.ko files on the host such that they can later be transferred and
loaded on the BBB.